Set server runtime when handling tunnel calls.
The client supplies its context to the server when making a mojit proxy (tunnel call) and in this context the runtime is set to client. The server has to change the context runtime to server when reusing this context, for example to ensure proper loading of configs specified as runtime:server.